Export html (#103)
add the `Experiment.export_html()` and its unit test;

update the html template, mostly changed to add shift+, shift_ to zoom in on sample axis and arrows for moving on sample axis as well. This because in the new chrome version, they disabled the shift+scroll wheel in d3.zoom object, so need to enable another way to zoom in.